Implemented correct (sequential) save to and load from Db
