+++ app/src/main/AndroidManifest.xml
@@ -24,22 +24,22 @@
         android:theme="@style/AppTheme">
         <activity android:name=".ui.activities.MainActivity"/>
         <activity android:name=".ui.activities.AuthActivity">
-            <!--<intent-filter>-->
-                <!--<action android:name="android.intent.action.MAIN"/>-->
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN"/>
 
-                <!--<category android:name="android.intent.category.LAUNCHER"/>-->
-            <!--</intent-filter>-->
+                <category android:name="android.intent.category.LAUNCHER"/>
+            </intent-filter>
         </activity>
         <activity android:name=".ui.activities.UserListActivity"
             android:label="DevIntesive Crew"
             >
 
-             <intent-filter>
-             <action android:name="android.intent.action.MAIN"/>
+             <!--<intent-filter>-->
+             <!--<action android:name="android.intent.action.MAIN"/>-->
 
 
-             <category android:name="android.intent.category.LAUNCHER"/>
-             </intent-filter>
+             <!--<category android:name="android.intent.category.LAUNCHER"/>-->
+             <!--</intent-filter>-->
         </activity>
         <activity android:name=".ui.activities.ProfileUserActivity" android:parentActivityName=".ui.activities.UserListActivity">
 <meta-data
