+++ app/src/main/java/com/softdesign/devintensive/utils/AppConfig.java
@@ -9,5 +9,5 @@ public interface AppConfig {
     String BASE_URL = "http://devintensive.softdesign-apps.ru/api/";
     int MAX_CONNECT_TIMEOUT = 10000;
     int MAX_READ_TIMEOUT = 10000;
-    int START_DELAY = 1500;
+    int START_DELAY = 1;
 }
